requirements.txt
asgiref==3.4.1
attrs==22.1.0
Django==3.2.15
django-cors-headers==3.10.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.4.0
drf-spectacular==0.24.2
importlib-metadata==4.8.3
inflection==0.5.1
jsonschema==3.2.0
Pillow==8.4.0
PyJWT==v1.7.1
pyrsistent==0.18.0
pytz==2022.4
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.1.1
uritemplate==4.1.1
zipp==3.6.0
